+++ components/vf-chatbot-dialog/CHANGELOG.md
@@ -3,3 +3,4 @@
 ### 1.0.0-beta.1
 
 * Added : Initial version for modal chatbot variant
+* Added : React support [Tracking issue](https://github.com/visual-framework/vf-core/issues/2314)